requirements_smtp.txt
flask>=2.0.0
flask-cors>=3.0.0
dnspython>=2.0.0
